// src/enum/errorType.enum.ts
export enum ErrorType {
  POST_NOT_FOUND = '게시물을 찾을 수 없습니다.',
  EXTERNAL_SNS_RESPONSE_FAILED = '외부 SNS API 응답에 실패하였습니다.',
  SNS_TYPE_NOT_FOUND = '타입을 찾을 수 없습니다.',
  PASSWORD_LENGTH_REQUIRE = '비밀번호는 10자 이상이여야 합니다.',
  PASSWORD_CHARACTER_REQUIRE = '비밀번호는 숫자, 문자, 특수문자 중 2가지 이상을 포함해야 합니다.',
  PASSWORD_DISALLOW_CONSECUTIVE = '비밀번호는 3회 이상 연속되는 문자 사용은 불가능합니다.',
  PASSWORD_MISMATCH = '비밀번호가 일치하지 않습니다.',
  PASSWORD_CONFIRM_MISMATCH = '비밀번호와 비밀번호 확인이 일치하지 않습니다.',
  EMAIL_NOT_VALID = '이메일 형식이 유효하지 않습니다.',
  USER_EXIST = '이미 존재하는 사용자입니다.',
  USER_NOT_EXIST = '사용자 정보가 존재하지 않습니다.',
  USER_NOT_APPROVE = '가입 승인되지 않은 사용자입니다.',
  USER_UNAUTHORIZED = '인증된 사용자가 아닙니다.',
  AUTH_CODE_INVALID = '유효하지 않은 인증코드 입니다.',
  AUTH_INVALID_TOKEN = '유효하지 않은 토큰입니다.',
  MAXIMUM_ALLOWED_DAYS_EXCEEDED = '최대 조회 가능한 날짜는 30일입니다.',
  MAXIMUM_ALLOWED_TIME_EXCEEDED = '최대 조회 가능한 시간은 일주일(168시간)입니다.',
}